requirements-test.txt
# =============================================================================
# ЗАВИСИМОСТИ ДЛЯ ТЕСТИРОВАНИЯ И РАЗРАБОТКИ
# =============================================================================
# Установка: pip install -r requirements.txt -r requirements-test.txt
#
# Примечание: Сначала установите основные зависимости из requirements.txt

# =============================================================================
# ТЕСТОВЫЕ ФРЕЙМВОРКИ
# =============================================================================
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
pytest-benchmark>=4.0.0
pytest-mock>=3.10.0

# =============================================================================
# НАГРУЗОЧНОЕ ТЕСТИРОВАНИЕ
# =============================================================================
locust>=2.14.0

# =============================================================================
# МОНИТОРИНГ ПРОИЗВОДИТЕЛЬНОСТИ
# =============================================================================
psutil>=5.9.0
memory-profiler>=0.60.0

# =============================================================================
# УТИЛИТЫ ДЛЯ ТЕСТОВ
# =============================================================================
factory-boy>=3.2.0
faker>=18.0.0
responses>=0.23.0
coverage>=7.0.0

# =============================================================================
# КАЧЕСТВО КОДА (опционально)
# =============================================================================
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
pylint>=2.15.0